requirements.txt
fastapi==0.111.0
redis==5.0.4
pydantic==2.9.2
pydantic_settings==2.6.0
uvicorn-worker==0.2.0
SQLAlchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
async-fastapi-jwt-auth==0.6.6
typer==0.14.0
passlib==1.7.4
Werkzeug==3.1.3
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
requests==2.32.3
